+++ templates/objc/header_file/___FILEBASENAME___.h
@@ -1,0 +1,13 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.
+//  ___COPYRIGHT___
+//
+
+#ifndef ___FILEBASENAME____h
+#define ___FILEBASENAME____h
+
+
+#endif /* ___FILEBASENAME____h */