+++ templates/objc/objc_file/category/___FILEBASENAME___.h
@@ -1,0 +1,17 @@
+//
+//  ___FILENAME___
+//  ___PROJECTNAME___
+//
+//  Created by ___FULLUSERNAME___ on ___DATE___.
+//  ___COPYRIGHT___
+//
+
+#import "___CLASSNAME___.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface ___CLASSNAME___ (___CATEGORYNAME___)
+
+@end
+
+NS_ASSUME_NONNULL_END